Update calculation.ts
Edited the code so that overlap in conditions for "De Hypersensitive Burn-out" and "De Chaotic Burn-out does not interfere with each other. 

Same with he condition for De Chaotic Burn-out and De Hypersensitive Burn-out. Doesn't consider both criteria simultaneously (focus and hypersensitivity). These should be handled individually, so I added that as well.